# docs/contracts/openapi/identity.yaml
openapi: 3.0.3
info:
  title: Vitalis Identity API
  version: 0.1.0
  description: |
    Autenticacao, usuarios, papeis e enderecos.

    Contrato canonico Vitalis-contracts. Clientes publicos via Gateway :8080.
servers:
  - url: http://localhost:8081
    description: Servico direto (rede interna)
  - url: http://localhost:8080
    description: Via Gateway
tags:
  - name: health
  - name: domain
security:
  - bearerAuth: []
paths:
  /health:
    get:
      tags: [health]
      summary: Liveness
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /ready:
    get:
      tags: [health]
      summary: Readiness
      security: []
      responses:
        '200': { description: Ready }
        '503': { description: Not ready }

  /v1/auth/login:
    post:
      tags: [domain]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, senha]
              properties:
                email: { type: string, format: email }
                senha: { type: string, format: password }
      responses:
        '200':
          description: Tokens emitidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, example: Bearer }
                  expires_in: { type: integer }
                  user:
                    type: object
                    properties:
                      id: { type: string }
                      nome: { type: string }
                      email: { type: string }
                      roles:
                        type: array
                        items: { type: string }
        '401':
          description: Credenciais invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/auth/refresh:
    post:
      tags: [domain]
      summary: Renovar access token (refresh rotacionado)
      security: []
      responses:
        '200': { description: Novo access token }
        '401': { description: Refresh invalido/revogado }
  /v1/auth/logout:
    post:
      tags: [domain]
      summary: Revoga refresh token
      responses:
        '204': { description: Logout OK }
  /v1/auth/registro:
    post:
      tags: [domain]
      summary: Cadastro de usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome, email, senha, tipo_usuario]
              properties:
                nome: { type: string }
                email: { type: string, format: email }
                senha: { type: string }
                telefone: { type: string }
                tipo_usuario:
                  type: string
                  enum: [paciente, medico, farmacia, motoboy, admin]
                cpf: { type: string }
                data_nascimento: { type: string, format: date }
      responses:
        '201': { description: Criado }
        '409': { description: Email ja existe }
  /v1/auth/esqueci-senha:
    post:
      tags: [domain]
      summary: Solicitar reset de senha
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202': { description: Aceito }
  /v1/auth/redefinir-senha:
    post:
      tags: [domain]
      summary: Redefinir senha com token
      security: []
      responses:
        '204': { description: Senha alterada }
  /v1/auth/jwks.json:
    get:
      tags: [domain]
      summary: JWKS publico RS256
      security: []
      responses:
        '200': { description: JWK Set }
  /v1/usuarios/me:
    get:
      tags: [domain]
      summary: Perfil do usuario autenticado
      responses:
        '200': { description: Perfil }
    put:
      tags: [domain]
      summary: Atualizar perfil
      responses:
        '200': { description: Atualizado }
  /v1/usuarios/me/enderecos:
    get:
      tags: [domain]
      summary: Listar enderecos
      responses:
        '200': { description: Lista }
    post:
      tags: [domain]
      summary: Criar endereco
      responses:
        '201': { description: Criado }
  /v1/usuarios/me/enderecos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [domain]
      summary: Atualizar endereco
      responses:
        '200': { description: OK }
    delete:
      tags: [domain]
      summary: Remover endereco
      responses:
        '204': { description: Removido }
  /v1/admin/usuarios/{id}/roles:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [domain]
      summary: Alterar papeis (admin)
      responses:
        '200': { description: Roles atualizadas }
        '403': { description: Sem permissao }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Obrigatorio em POST de cobranca/assinatura (Billing).
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema: { type: string }
  schemas:
    Error:
      type: object
      required: [erro, codigo]
      properties:
        erro: { type: string }
        codigo: { type: string }
        request_id: { type: string }
        detalhes:
          type: object
          additionalProperties: true
    Health:
      type: object
      properties:
        status: { type: string, enum: [ok, degraded] }
        service: { type: string }